requirements.txt
altair==5.5.0
asttokens @ file:///croot/asttokens_1743630435401/work
attrs @ file:///home/task_176235647440041/conda-bld/attrs_1762356877723/work
black==25.11.0
blinker==1.9.0
Bottleneck @ file:///home/task_176193794916414/conda-bld/bottleneck_1761937997056/work
cachetools==6.2.2
certifi==2025.11.12
cffi @ file:///home/task_176183215096628/conda-bld/cffi_1761832773152/work
charset-normalizer==3.4.4
click==8.3.1
comm @ file:///home/task_176311912503377/conda-bld/comm_1763119154572/work
confluent-kafka==2.12.2
cryptography @ file:///home/task_176193171464307/conda-bld/cryptography-split_1761931963077/work
debugpy @ file:///home/task_176242086745011/conda-bld/debugpy_1762421636577/work
decorator @ file:///home/task_175734099691499/conda-bld/decorator_1757341236139/work
dnspython==1.16.0
executing @ file:///home/task_175706097783255/conda-bld/executing_1757061230504/work
flake8==7.3.0
gitdb==4.0.12
GitPython==3.1.45
idna @ file:///home/task_176191177554882/conda-bld/idna_1761911981359/work
iniconfig==2.3.0
ipykernel @ file:///home/task_176294602487836/conda-bld/ipykernel_1762946074244/work
ipython @ file:///home/task_176278907235969/conda-bld/ipython_1762789338226/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.6
jsonschema @ file:///home/task_176245919820156/conda-bld/jsonschema_1762459245784/work
jsonschema-specifications @ file:///home/task_176278841581006/conda-bld/jsonschema-specifications_1762788429133/work
jupyter_client @ file:///home/task_176294303581930/conda-bld/jupyter_client_1762943085155/work
jupyter_core @ file:///croot/jupyter_core_1751991368470/work
kafka-python==2.3.0
MarkupSafe==3.0.3
matplotlib-inline @ file:///home/task_176277916602646/conda-bld/matplotlib-inline_1762779180318/work
mccabe==0.7.0
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159252635933/conda-bld/mkl_fft_1761592902019/work
mkl_random @ file:///home/task_176159299715632/conda-bld/mkl_random_1761593011445/work
mypy_extensions==1.1.0
narwhals==2.12.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
numexpr @ file:///home/task_176216555051259/conda-bld/numexpr_1762165561927/work
numpy @ file:///home/task_176252831629565/conda-bld/numpy_and_numpy_base_1762528499256/work/dist/numpy-2.3.4-cp312-cp312-linux_x86_64.whl#sha256=7c8e1bc25ef710e8a712e0d3f02dfc4501bf9da0e0d91b55b7519037a5902dfb
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
pandas @ file:///home/task_176233227348745/conda-bld/pandas_1762332326353/work/dist/pandas-2.3.3-cp312-cp312-linux_x86_64.whl#sha256=0864c010797a6bd32dd751b2e8470cd1061bae036d65da48dc69a4606b4674eb
parso @ file:///home/task_176278181429383/conda-bld/parso_1762781859742/work
pathspec==0.12.1
pexpect @ file:///home/task_176253513024890/conda-bld/pexpect_1762535937809/work
pillow==12.0.0
platformdirs @ file:///home/task_176235638281825/conda-bld/platformdirs_1762356487196/work
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit @ file:///home/task_176174483920045/conda-bld/prompt-toolkit_1761744853185/work
protobuf==6.33.1
psutil @ file:///home/task_176189642211680/conda-bld/psutil_1761896463664/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///home/task_175706684826776/conda-bld/pure_eval_1757067049051/work
py4j @ file:///home/task_175813273798106/conda-bld/py4j_1758132799593/work
pyarrow @ file:///home/task_175983286518928/conda-bld/pyarrow_1759833585098/work/python
pycodestyle==2.14.0
pycparser @ file:///home/task_175749564075091/conda-bld/pycparser_1757496039511/work
pydeck==0.9.1
pyflakes==3.4.0
Pygments @ file:///home/task_176243120367123/conda-bld/pygments_1762431407417/work
pymongo==3.12.0
pyspark @ file:///home/task_175829495072888/conda-bld/pyspark_1758295010548/work
pytest==9.0.1
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.1
pytokens==0.3.0
pytz @ file:///croot/pytz_1752135852232/work
pyzmq @ file:///home/task_176237559504765/conda-bld/pyzmq_1762375608075/work
referencing @ file:///home/task_176253675477096/conda-bld/referencing_1762536887597/work
requests==2.32.5
rpds-py @ file:///home/task_176236637726140/conda-bld/rpds-py_1762366483045/work
setuptools==80.9.0
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
stack_data @ file:///home/task_175706679004827/conda-bld/stack_data_1757067036633/work
streamlit==1.51.0
streamlit-autorefresh==1.0.1
tenacity==9.1.2
toml==0.10.2
tornado @ file:///croot/tornado_1748956929273/work
traitlets @ file:///croot/traitlets_1718227057033/work
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.5.0
watchdog==6.0.0
wcwidth @ file:///croot/wcwidth_1750352883074/work
wheel==0.45.1
